OLMIS-545 Update docker compose to add dependency on reference data
Add entry for reference data. Instead of using links, use a custom network configuration that all containers will be a part of. This configuration doesn't work in a dev environment because of a circular dependency (auth depends on reference data for external calls, but reference data also depends on auth for authentication). There are multiple issues:

* auth and/or reference data cannot run b/c db may not be up by the time they need to access it. Workaround is to run db first (docker-compose up db)
* cannot run auth in dev mode b/c then reference data, which is started up before it, will not know about it in the network. To get it to work at all, must run rest of the services in non-dev mode (docker-compose -f docker-compose.yml up log referencedata auth). However, this does not allow incremental dev changes to auth.