OLMIS-1930 Changed forgot password email template
Change template format to be a bit more readable. Could not include first and last name because auth does not have that info stored in User. Either adding those fields into auth or calling reference data for it would require more LoE than we have time for this sprint/3.0 release.