upgrade to dev image v3
moving to Gradle v4 from v2.13 also meant that:
- sonarqube plugin had to be updated
- flyway and node plugins could be updated
- the generate migration task was using gradle's "left shift" function
which became deprecated.  Using this function doesn't seem necessary
so it was converted into a regular task.
- it does appear that even with the latest flyway gradle plugin, that
it might still use deprecated gradle functions.  For now this seems
harmless though a warning will be given, and flyway's plugin should
be updated when a new one is available.